Move some colors into core-design-system
