+++ app/src/main/java/com/hedvig/app/feature/claimdetail/ui/FakeWaveAudioPlayerCard.kt
@@ -40,7 +40,7 @@ import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.app.R
 import com.hedvig.app.service.audioplayer.AudioPlayerState
 import com.hedvig.app.service.audioplayer.AudioPlayerState.Ready.ReadyState
-import com.hedvig.app.ui.compose.theme.onWarning
+import com.hedvig.android.core.designsystem.theme.onWarning
 import com.hedvig.app.ui.compose.theme.warning
 import com.hedvig.app.util.ProgressPercentage
