+++ app/src/main/java/com/hedvig/app/feature/insurance/ui/CrossSell.kt
@@ -32,9 +32,9 @@ import coil.compose.rememberImagePainter
 import com.commit451.coiltransformations.CropTransformation
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.app.feature.crossselling.ui.CrossSellData
-import com.hedvig.app.ui.compose.theme.hedvigBlack
-import com.hedvig.app.ui.compose.theme.hedvigBlack12percent
-import com.hedvig.app.ui.compose.theme.whiteHighEmphasis
+import com.hedvig.android.core.designsystem.theme.hedvigBlack
+import com.hedvig.android.core.designsystem.theme.hedvigBlack12percent
+import com.hedvig.android.core.designsystem.theme.whiteHighEmphasis
 import com.hedvig.app.util.compose.rememberBlurHash
 
 /*
