+++ app/src/main/java/com/hedvig/app/feature/offer/ui/composable/insurely/InsurelyCard.kt
@@ -20,7 +20,7 @@ import com.hedvig.app.feature.offer.ui.OfferItems
 import com.hedvig.app.feature.offer.ui.OfferItems.InsurelyCard.FailedToRetrieve
 import com.hedvig.app.feature.offer.ui.OfferItems.InsurelyCard.Loading
 import com.hedvig.app.feature.offer.ui.OfferItems.InsurelyCard.Retrieved
-import com.hedvig.app.ui.compose.theme.hedvigBlack12percent
+import com.hedvig.android.core.designsystem.theme.hedvigBlack12percent
 import com.hedvig.app.ui.compose.theme.hedvigContentColorFor
 import com.hedvig.app.util.compose.preview.previewData
 import java.util.Locale
