+++ app/src/main/java/com/hedvig/app/ui/compose/composables/banner/InfoBanner.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
 import com.hedvig.app.R
-import com.hedvig.app.ui.compose.theme.hedvigBlack12percent
+import com.hedvig.android.core.designsystem.theme.hedvigBlack12percent
 
 @Composable
 fun InfoBanner(
