+++ app/src/main/java/com/hedvig/app/ui/compose/theme/Colors.kt
@@ -9,25 +9,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
 import com.hedvig.app.R
 
-val hedvigBlack = Color(0xff121212)
-val hedvigBlack12percent = hedvigBlack.copy(alpha = 0.12f)
-val hedvigOffWhite = Color(0xfffafafa)
-val hedvigDarkGray = Color(0xff505050)
-val background = Color(0xffF6F6F6)
-
-val whiteHighEmphasis = Color(0xFFFAFAFA)
-
-val errorLight = Color(0xffDD2727)
-val errorDark = Color(0xffE24646)
-
-val textColorPrimary = Color(0xAB121212)
-val textColorPrimaryDark = Color(0x8FFAFAFA)
-
-val surfaceDark = Color(0xffBE9BF3)
-
-val progressBlue = Color(0xffC3CBD6)
-val progressYellow = Color(0xffEDCDAB)
-
 @Composable
 fun hedvigContentColorFor(backgroundColor: Color): Color {
   return when (backgroundColor) {
@@ -45,16 +26,3 @@ val Colors.warning: Color
   @Composable
   @ReadOnlyComposable
   get() = colorResource(R.color.colorWarning)
-
-@Suppress("unused")
-val Colors.onWarning: Color
-  @Composable
-  @ReadOnlyComposable
-  get() = colorResource(R.color.hedvig_black)
-
-val Colors.separator: Color
-  get() = if (isLight) {
-    hedvigBlack.copy(alpha = 0.12f)
-  } else {
-    hedvigOffWhite.copy(alpha = 0.12f)
-  }
